requirements.txt

Django==2.0.8
django-crispy-forms==1.7.2
Pillow==5.4.1